migrate scope fields to scope module
